Change Float to Double

+++ src/main/java/com/hackclub/hccore/listeners/AdvancementListener.java
@@ -69,7 +69,7 @@ public class AdvancementListener implements Listener {
         Player player = (Player) event.getEntity();
 
         // Check that the value is equal or above that of 1 million miles (1,609,344 km)
-        final float METERS_PER_MILE = 1609.344;
+        final double METERS_PER_MILE = 1609.344;
         if (player.getStatistic(Statistic.FLY_ONE_CM) <= (1000000 * METERS_PER_MILE)) {
             return;
         }
